Refactor: Move handleShowReviews to ReviewButton
