+++ package.json
@@ -22,7 +22,7 @@
     "react-native-gesture-handler": "^2.20.2",
     "react-native-image-picker": "^7.2.3",
     "react-native-keyboard-aware-scroll-view": "^0.9.5",
-    "react-native-permissions": "^5.2.2",
+    "react-native-permissions": "^5.2.3",
     "react-native-reanimated": "^3.16.6",
     "react-native-safe-area-context": "^4.14.0",
     "react-native-screens": "^4.0.0",
